Update: カウントとIDの初期値指定を上位もコンポーネントに切り出し
